Migration to OpenVox and Vagrantfile enhancements

+++ Vagrantfile
@@ -1,55 +1,70 @@
+# frozen_string_literal: true
+
 # -*- mode: ruby -*-
 # vi: set ft=ruby :
 
-INSTANCES=2
+INSTANCES = 2
 
-PROVISION_PUPPET = <<PUPPET
-/bin/rpm -Uvh https://yum.puppet.com/puppet7-release-el-8.noarch.rpm
-/usr/bin/dnf -y install puppet-agent
-echo '*' > /etc/puppetlabs/puppet/autosign.conf
-/opt/puppetlabs/bin/puppet resource host puppet.choria ensure=present ip=192.168.56.5 host_aliases=puppet
-mkdir -p /etc/puppetlabs/facter/facts.d
-echo "role=${1}" > /etc/puppetlabs/facter/facts.d/role.txt
-PUPPET
+def read_box_name
+  config = JSON.parse(File.read(File.join(Dir.pwd, 'automation_config.json')))
+  config['boxName']
+end
 
-Vagrant.configure("2") do |config|
-  # Monta o projeto inteiro em /vagrant (padrão do Vagrant)
-  config.vm.synced_folder ".", "/vagrant", type: "virtualbox"
+BOX_NAME = read_box_name
+
+PROVISION_OPENVOX = <<~OPENVOX
+  ROLE_FILE=/etc/puppetlabs/facter/facts.d/role.txt
+  if [[ -f $ROLE_FILE ]]
+  then
+      echo 'Packages already installed, nothing to do'
+  else
+      echo '*' > /etc/puppetlabs/puppet/autosign.conf
+      /opt/puppetlabs/bin/puppet resource host puppet.choria ensure=present ip=192.168.56.5 host_aliases=puppet
+      mkdir -p /etc/puppetlabs/facter/facts.d
+      echo "role=${1}" > /etc/puppetlabs/facter/facts.d/role.txt
+  fi
+OPENVOX
+
+Vagrant.configure('2') do |config|
+  config.vm.synced_folder '.', '/vagrant', type: 'virtualbox'
 
   config.vm.define :puppet do |vmconfig|
-    vmconfig.vm.box = "generic/rocky8"
-    vmconfig.vm.hostname = "puppet.choria"
-    vmconfig.vm.network :private_network, ip: "192.168.56.5"
+    vmconfig.vm.box = BOX_NAME
+    vmconfig.vm.hostname = 'puppet.choria'
+    vmconfig.vm.network :private_network, ip: '192.168.56.5'
+    vmconfig.vm.network :forwarded_port, guest: 9100, host: 9100, id: 'Prometheus'
     vmconfig.vm.provider :virtualbox do |vb|
-        vb.customize ["modifyvm", :id, "--memory", 3072]
+      vb.linked_clone = true
+      vb.customize ['modifyvm', :id, '--memory', 3072]
+      vb.customize ['modifyvm', :id, '--graphicscontroller', 'vmsvga']
+      vb.customize ['modifyvm', :id, '--vram', '16']
+      vb.name = "openvox"
     end
 
-    vmconfig.vbguest.installer_options = { allow_kernel_upgrade: true }
     vmconfig.vbguest.auto_update = false
 
     vmconfig.vm.provision :shell do |s|
-      s.inline = PROVISION_PUPPET
-      s.args = "puppetserver"
+      s.inline = PROVISION_OPENVOX
+      s.args = 'puppetserver'
     end
 
-    vmconfig.vm.provision "shell", inline: <<-SHELL
-      echo ">>> Installing PuppetServer..."
-      /bin/rpm -Uvh https://yum.puppet.com/puppetserver-release-el-8.noarch.rpm
-      dnf install -y puppetserver
-
-      echo ">>> Syncing environments to /etc/puppetlabs/code/environments..."
-      rsync -a --delete /vagrant/environments/ /etc/puppetlabs/code/environments/
-
-      echo ">>> Starting PuppetServer..."
-      systemctl start puppetserver
-      systemctl enable puppetserver
-
-      sleep 10
-      systemctl status puppetserver --no-pager | head -5
+    vmconfig.vm.provision 'shell', inline: <<-SHELL
+      SERVICE_NAME=puppetserver
+      if systemctl status $SERVICE_NAME --no-pager &> /dev/null
+      then
+          echo 'OpenVox already installed, nothing to do'
+      else
+          /usr/bin/dnf install -y openvox-server
+          echo ">>> Syncing environments to /etc/puppetlabs/code/environments..."
+          rsync -a --delete /vagrant/environments/ /etc/puppetlabs/code/environments/
+          systemctl start $SERVICE_NAME
+          systemctl enable $SERVICE_NAME
+          sleep 10
+          systemctl status $SERVICE_NAME --no-pager | head -5
+      fi
     SHELL
 
-    vmconfig.vm.provision "shell", inline: <<-SHELL
-      echo ">>> Running puppet apply to configure Choria and Puppet..."
+    vmconfig.vm.provision 'shell', inline: <<-SHELL
       /opt/puppetlabs/bin/puppet apply /etc/puppetlabs/code/environments/production/manifests/default.pp \
         --hiera_config=/etc/puppetlabs/code/environments/production/hiera.yaml \
         --modulepath=/etc/puppetlabs/code/environments/production/modules:/etc/puppetlabs/code/environments/production/site
@@ -58,22 +73,25 @@ Vagrant.configure("2") do |config|
 
   INSTANCES.times do |i|
     config.vm.define "instance#{i}" do |vmconfig|
-      vmconfig.vm.box = "generic/rocky8"
-      vmconfig.vm.hostname = "choria%s.choria" % i
-      vmconfig.vm.network :private_network, ip: "192.168.56.%d" % (9+i)
+      vmconfig.vm.box = BOX_NAME
+      vmconfig.vm.hostname = "choria#{i}.choria"
+      vmconfig.vm.network :private_network, ip: format('192.168.56.%d', 9 + i)
       vmconfig.vm.provider :virtualbox do |vb|
-          vb.customize ["modifyvm", :id, "--memory", 1024]
+        vb.customize ['modifyvm', :id, '--memory', 1024]
+        vb.customize ['modifyvm', :id, '--graphicscontroller', 'vmsvga']
+        vb.customize ['modifyvm', :id, '--vram', '16']
+        vb.linked_clone = true
+        vb.name = "instance#{i}"
       end
 
-      vmconfig.vbguest.installer_options = { allow_kernel_upgrade: true }
       vmconfig.vbguest.auto_update = false
 
       vmconfig.vm.provision :shell do |s|
-        s.inline = PROVISION_PUPPET
-        s.args = "managed"
+        s.inline = PROVISION_OPENVOX
+        s.args = 'managed'
       end
 
-      vmconfig.vm.provision "shell", inline: <<-SHELL
+      vmconfig.vm.provision 'shell', inline: <<-SHELL
         echo ">>> Syncing environments to /etc/puppetlabs/code/environments..."
         rsync -a --delete /vagrant/environments/ /etc/puppetlabs/code/environments/
 
@@ -82,7 +100,8 @@ Vagrant.configure("2") do |config|
           --hiera_config=/etc/puppetlabs/code/environments/production/hiera.yaml \
           --modulepath=/etc/puppetlabs/code/environments/production/modules:/etc/puppetlabs/code/environments/production/site
       SHELL
-      vmconfig.vm.provision "shell", inline: <<-SHELL
+
+      vmconfig.vm.provision 'shell', inline: <<-SHELL
         echo ">>> Running puppet agent to enroll with PuppetServer..."
         /opt/puppetlabs/bin/puppet agent -tv --waitforcert 30 --server puppet.choria
       SHELL
